requirements.txt
tabulate>=0.9.0
httpx>=0.24.1
beautifulsoup4>=4.12.0
pandas>=2.0.0
requests>=2.28.0
numpy>=1.22.0
python-dateutil>=2.8.2
pytest>=7.0.0
asyncio>=3.4.3
PyPDF2>=3.0.0
pdfminer.six>=20221105
streamlit>=1.18.0
fake-useragent>=1.1.3
lxml>=4.9.2
selenium>=4.10.0
webdriver-manager>=3.8.6
sentence-transformers>=2.2.2
# Dépendances pour le scraping
apache-airflow>=2.6.0
apscheduler>=3.10.0 